feat: add versioning to bytecode programs and improve error handling for host calls
