+++ pkg/bytecode/program.go
@@ -19,6 +19,7 @@ type (
 	}
 
 	Program struct {
+		Version    int
 		Source     *file.Source
 		Registers  int
 		Bytecode   []Instruction
@@ -29,6 +30,8 @@ type (
 		Metadata   Metadata
 	}
 )
+
+const ProgramVersion = 2
 
 func (p *Program) MarshalJSON() ([]byte, error) {
 	if p == nil {
@@ -48,6 +51,7 @@ func (p *Program) MarshalJSON() ([]byte, error) {
 	}
 
 	payload := programJSON{
+		Version:    p.Version,
 		Source:     p.Source,
 		Registers:  p.Registers,
 		Bytecode:   p.Bytecode,
@@ -84,6 +88,7 @@ func (p *Program) UnmarshalJSON(data []byte) error {
 	}
 
 	p.Source = decoded.Source
+	p.Version = decoded.Version
 	p.Registers = decoded.Registers
 	p.Bytecode = decoded.Bytecode
 	p.Constants = constants
