+++ pkg/vm/internal/data/collector_key_counter.go
@@ -45,35 +45,13 @@ func (c *KeyCounterCollector) Iterate(ctx context.Context) (runtime.Iterator, er
 }
 
 func (c *KeyCounterCollector) sort(ctx context.Context) error {
-	return runtime.SortListWith(ctx, c.Value, func(first, second runtime.Value) int {
-		firstKV, firstOk := first.(*KV)
-		secondKV, secondOk := second.(*KV)
-
-		var comp int
-
-		if firstOk && secondOk {
-			comp = runtime.CompareValues(firstKV.Key, secondKV.Key)
-		} else {
-			comp = runtime.CompareValues(first, second)
-		}
-
-		return comp
-	})
+	return sortCollectorList(ctx, c.Value)
 }
 
 func (c *KeyCounterCollector) Set(ctx context.Context, key, _ runtime.Value) error {
-	var keyStr string
-
-	switch k := key.(type) {
-	case runtime.String:
-		keyStr = k.String()
-	default:
-		var err error
-		keyStr, err = Stringify(ctx, key)
-
-		if err != nil {
-			return err
-		}
+	keyStr, err := normalizeCollectorKey(ctx, key)
+	if err != nil {
+		return err
 	}
 
 	// Fast path: first key stays in singleKey/singleKV to avoid map allocation.
@@ -163,18 +141,9 @@ func (c *KeyCounterCollector) Set(ctx context.Context, key, _ runtime.Value) err
 }
 
 func (c *KeyCounterCollector) Get(ctx context.Context, key runtime.Value) (runtime.Value, error) {
-	var keyStr string
-
-	switch k := key.(type) {
-	case runtime.String:
-		keyStr = k.String()
-	default:
-		var err error
-		keyStr, err = Stringify(ctx, key)
-
-		if err != nil {
-			return nil, err
-		}
+	keyStr, err := normalizeCollectorKey(ctx, key)
+	if err != nil {
+		return nil, err
 	}
 
 	if c.grouping == nil {
@@ -182,13 +151,13 @@ func (c *KeyCounterCollector) Get(ctx context.Context, key runtime.Value) (runti
 			return c.singleIndex, nil
 		}
 
-		return runtime.None, runtime.Errorf(runtime.ErrNotFound, "collector key: %s", keyStr)
+		return runtime.None, collectorKeyNotFound(keyStr)
 	}
 
 	v, ok := c.grouping[keyStr]
 
 	if !ok {
-		return runtime.None, runtime.Errorf(runtime.ErrNotFound, "collector key: %s", keyStr)
+		return runtime.None, collectorKeyNotFound(keyStr)
 	}
 
 	return v, nil
