+++ test/integration/optimization/assertions.go
@@ -177,32 +177,6 @@ func instructionUseDef(inst bytecode.Instruction) (uses []int, defs []int) {
 			defs = append(defs, op.Register())
 		}
 	}
-	addRangeUses := func(start, end bytecode.Operand) {
-		if !start.IsRegister() || !end.IsRegister() {
-			return
-		}
-		startReg := start.Register()
-		endReg := end.Register()
-		if startReg <= 0 || endReg < startReg {
-			return
-		}
-		for r := startReg; r <= endReg; r++ {
-			uses = append(uses, r)
-		}
-	}
-	addFixedRangeUses := func(start bytecode.Operand, count int) {
-		if !start.IsRegister() {
-			return
-		}
-		startReg := start.Register()
-		if startReg <= 0 || count <= 0 {
-			return
-		}
-		for r := startReg; r < startReg+count; r++ {
-			uses = append(uses, r)
-		}
-	}
-
 	op := inst.Opcode
 	dst, src1, src2 := inst.Operands[0], inst.Operands[1], inst.Operands[2]
 
@@ -297,71 +271,14 @@ func instructionUseDef(inst bytecode.Instruction) (uses []int, defs []int) {
 		return
 
 	// Calls.
-	case bytecode.OpHCall, bytecode.OpProtectedHCall:
-		addRangeUses(src1, src2)
-		addDef(dst)
-		return
-	case bytecode.OpHCall0, bytecode.OpProtectedHCall0:
-		addDef(dst)
-		return
-	case bytecode.OpHCall1, bytecode.OpProtectedHCall1:
-		addUse(src1)
-		addDef(dst)
-		return
-	case bytecode.OpHCall2, bytecode.OpProtectedHCall2:
-		addUse(src1)
-		addUse(src2)
-		addDef(dst)
-		return
-	case bytecode.OpHCall3, bytecode.OpProtectedHCall3:
-		addFixedRangeUses(src1, 3)
-		addDef(dst)
-		return
-	case bytecode.OpHCall4, bytecode.OpProtectedHCall4:
-		addFixedRangeUses(src1, 4)
-		addDef(dst)
-		return
-	case bytecode.OpCall, bytecode.OpProtectedCall:
-		addRangeUses(src1, src2)
-		addDef(dst)
-		return
-	case bytecode.OpCall0, bytecode.OpProtectedCall0:
-		addDef(dst)
-		return
-	case bytecode.OpCall1, bytecode.OpProtectedCall1:
-		addUse(src1)
-		addDef(dst)
-		return
-	case bytecode.OpCall2, bytecode.OpProtectedCall2:
-		addUse(src1)
-		addUse(src2)
-		addDef(dst)
-		return
-	case bytecode.OpCall3, bytecode.OpProtectedCall3:
-		addFixedRangeUses(src1, 3)
-		addDef(dst)
-		return
-	case bytecode.OpCall4, bytecode.OpProtectedCall4:
-		addFixedRangeUses(src1, 4)
-		addDef(dst)
-		return
-	case bytecode.OpTailCall:
-		addRangeUses(src1, src2)
-		return
-	case bytecode.OpTailCall0:
-		return
-	case bytecode.OpTailCall1:
-		addUse(src1)
-		return
-	case bytecode.OpTailCall2:
-		addUse(src1)
-		addUse(src2)
-		return
-	case bytecode.OpTailCall3:
-		addFixedRangeUses(src1, 3)
-		return
-	case bytecode.OpTailCall4:
-		addFixedRangeUses(src1, 4)
+	case bytecode.OpHCall, bytecode.OpProtectedHCall, bytecode.OpCall, bytecode.OpProtectedCall, bytecode.OpTailCall:
+		addUse(dst)
+		bytecode.VisitCallArgumentRegisters(op, src1, src2, func(reg int) {
+			uses = append(uses, reg)
+		})
+		if op != bytecode.OpTailCall {
+			addDef(dst)
+		}
 		return
 
 	// Stream.
